// src/app/model/gestionCobranza/Clientes_morosidad_ext.ts
export class Clientes_morosidad_ext {
  id_cliente: number = 0;
  nombre_completo: string = '';
  tipodocident: number = 0;
  nrodocident: string = '';
  desctipo: string = '';
  id_contrato: number = 0;
  deuda_actual_soles: number = 0;
  dias_en_mora: number = 0;
  ultimo_pago: string = '';
  estado: string = '';
  facturas_en_mora: number = 0;
  telefono: string = '';
  email: string = '';
}