// src/app/model/gestionCobranza/Detalle_facturas_mora.ts
export class Detalle_facturas_mora {
  id_factura: number = 0;
  id_contrato: number = 0;
  codigo_factura: string = '';
  tipo_servicio: string = '';
  periodo: string = '';
  fecha_emisión: string = '';
  fecha_vencimiento: string = '';
  monto: number = 0;
  saldo: number = 0;
  dias_mora: number = 0;
  estado: string = '';
}